Filtering logic from get all bootcamps is extracted
